refactor(routes): fiat deposit and withdrawal route names
