+++ src/app/fiat-deposit/[exchangeName]/query-func.ts
