+++ src/app/fiat-withdrawal/page.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { Heading, VStack } from "@/src/components/chakra";
-import Fiat from "@/src/app/deposit/fiat";
+import Fiat from "@/src/app/fiat-withdrawal/fiat";
 import { getLastYearTimestamp } from "@/src/utils/date";
 export default async function Deposit() {
   return (
     <VStack align={"flex-start"} spacing={4}>
-      <Heading>Deposits</Heading>
+      <Heading>Withdrawals</Heading>
       <Fiat timestamp={getLastYearTimestamp()} />
     </VStack>
   );
